backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
python-dotenv==1.0.1
jinja2==3.1.4
reportlab==4.2.2
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
sqlalchemy==2.0.36
alembic==1.13.3
python-multipart==0.0.9
pytz==2024.2
openpyxl==3.1.5
apscheduler==3.10.4
pytest==8.3.3
requests==2.32.3
python-dateutil==2.9.0
piper-tts>=1.3.0
